# cmd/api/docs/swagger.yaml
definitions:
  entity.User:
    properties:
      id:
        type: integer
      name:
        type: string
      surname:
        type: string
    type: object
  handler.Response:
    properties:
      id:
        type: integer
      name:
        type: string
      surname:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      description: Get all users
      operationId: FindAll
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Response'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user
      operationId: Create
      parameters:
      - description: entity.User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Modify a user
      operationId: Modify
      parameters:
      - description: entity.User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Modify a user
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Delete a user
      operationId: Delete
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      description: Get a user by ID
      operationId: FindByID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a user by ID
      tags:
      - users
swagger: "2.0"